backend/requirements.txt
blinker==1.8.2
click==8.1.7
confluent-kafka==2.4.0
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
dnspython==2.6.1
et-xmlfile==1.1.0
Flask==3.0.3
fonttools==4.53.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
numpy==1.26.4
openpyxl==3.1.3
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
pmdarima==2.0.4
pymongo==4.7.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.0
scipy==1.13.1
statsmodels==0.14.2
threadpoolctl==3.5.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
xlrd==2.0.1
zipp==3.19.2
